refactor: MyChartの週次データ生成ロジックをutilsに分離。理由: ロジックの共通化と再利用性向上のため。
